Changed User model, added AbstractUser
