+++ project/api/admin.py
@@ -1,7 +1,6 @@
 from django import forms
 from django.contrib import admin
 from django.contrib.auth import get_user_model
-from django.contrib.auth.admin import UserAdmin
 from django.utils.translation import gettext_lazy as _
 
 from . import models
@@ -12,41 +11,6 @@ User = get_user_model()
 
 class MixinAdmin(admin.ModelAdmin):
     empty_value_display = _('-пусто-')
-
-
-admin.site.unregister(User)
-
-
-@admin.register(User)
-class CustomUserAdmin(UserAdmin):
-    readonly_fields = ('date_joined', 'last_login')
-
-    def get_form(self, request, obj=None, **kwargs):
-        form = super().get_form(request, obj, **kwargs)
-        is_superuser = request.user.is_superuser
-        disabled_fields = set()
-        if not is_superuser:
-            disabled_fields |= {
-                'username',
-                'is_superuser',
-                'user_permissions',
-            }
-        if (
-            not is_superuser
-            and obj is not None
-            and (obj.is_superuser or obj == request.user)
-        ):
-            disabled_fields |= {
-                'is_active',
-                'is_staff',
-                'is_superuser',
-                'groups',
-                'user_permissions',
-            }
-        for f in disabled_fields:
-            if f in form.base_fields:
-                form.base_fields[f].disabled = True
-        return form
 
 
 @admin.register(models.Article)
@@ -77,8 +41,7 @@ class EventAdmin(MixinAdmin):
         queryset = super().get_queryset(request)
         if request.user.has_perm('api.view_all_cities'):
             return queryset
-        profile = request.user.profile
-        return queryset.filter(city__in=profile.region.cities.all())
+        return queryset.filter(city__in=request.user.region.cities.all())
 
 
 @admin.register(models.History)
@@ -121,14 +84,6 @@ class ParticipantAdmin(MixinAdmin):
     list_display = ('id', 'event', 'participant')
     search_fields = ('event', 'participant')
     autocomplete_fields = ('event', 'participant')
-
-
-@admin.register(models.Profile)
-class ProfileAdmin(MixinAdmin):
-    list_display = ('id', 'user', 'city', 'region')
-    search_fields = ('user', 'city', 'region')
-    list_filter = ('city', 'region')
-    autocomplete_fields = ('city', 'region')
 
 
 @admin.register(models.Region)
