+++ project/api/serializers/profile.py
@@ -1,26 +1,10 @@
+from django.contrib.auth import get_user_model
 from rest_framework import serializers
 
-from ..models import Profile
+User = get_user_model()
 
 
 class ProfileSerializer(serializers.ModelSerializer):
-    username = serializers.CharField(source='user.username')
-    first_name = serializers.CharField(source='user.first_name')
-    last_name = serializers.CharField(source='user.last_name')
-    email = serializers.EmailField(source='user.email')
-
     class Meta:
         fields = ['username', 'first_name', 'last_name', 'email', 'city']
-        model = Profile
-
-    def update(self, instance, validated_data):
-        user_data = validated_data.pop('user', {})
-        user = instance.user
-
-        user.username = user_data.get('username', user.username)
-        user.first_name = user_data.get('first_name', user.first_name)
-        user.last_name = user_data.get('last_name', user.last_name)
-        user.email = user_data.get('email', user.email)
-
-        super(ProfileSerializer, self).update(instance, validated_data)
-        return instance
+        model = User
