+++ project/api/views/event.py
@@ -39,7 +39,7 @@ class EventViewSet(viewsets.ViewSet):
             return queryset.order_by('start_at')
         booked = Event.objects.filter(pk=OuterRef('pk'), participants=user)
         queryset = queryset.annotate(
-            booked=Exists(booked)).filter(city=user.profile.city)
+            booked=Exists(booked)).filter(city=user.city)
         return queryset.order_by('start_at')
 
 
@@ -52,9 +52,9 @@ class ParticipantViewSet(ListCreateDelViewSet):
         return Participant.objects.filter(participant=self.request.user)
 
     def create(self, request):
-        event = get_object_or_404(Event, id=self.request.data.get('event'))
+        event = get_object_or_404(Event, id=self.request.POST.get('event'))
         self.check_object_permissions(self.request, event)
-        serializer = self.get_serializer(data=self.request.data)
+        serializer = self.get_serializer(data=self.request.POST)
         serializer.is_valid(raise_exception=True)
         self.perform_create(serializer)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
