+++ project/api/views/profile.py
@@ -1,21 +1,16 @@
 from django.contrib.auth import get_user_model
-from django.shortcuts import get_object_or_404
 from rest_framework.generics import RetrieveUpdateAPIView
 from rest_framework.permissions import IsAuthenticated
 
-from ..models import Profile
 from ..serializers import ProfileSerializer
 
 User = get_user_model()
 
 
 class ProfileViewSet(RetrieveUpdateAPIView):
-    queryset = Profile.objects.all()
+    queryset = User.objects.all()
     serializer_class = ProfileSerializer
     permission_classes = [IsAuthenticated]
 
     def get_object(self):
-        queryset = self.get_queryset()
-        obj = get_object_or_404(queryset, user=self.request.user)
-        self.check_object_permissions(self.request, obj)
-        return obj
+        return self.request.user
